delete html
