--- chara.html
@@ -1,39 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-    <head>
-        <meta charset="UTF-8">
-        <meta name="viewport" content="width=device-width, initial-scale=1.0">
-        <title>Player Creator</title>
-        <link rel="stylesheet" href="./assets/style/style.css"/>
-        <link rel="stylesheet" href="./assets/style/index.css"/>
-        <link rel="preconnect" href="https://fonts.googleapis.com">
-        <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-        <link href="https://fonts.googleapis.com/css2?family=Cinzel+Decorative&display=swap" rel="stylesheet">
-        <link href="https://fonts.googleapis.com/css2?family=Raleway:wght@500&display=swap" rel="stylesheet">
-        <link href="https://fonts.googleapis.com/css2?family=Fondamento&display=swap" rel="stylesheet">
-        <link href="https://fonts.googleapis.com/css2?family=Amita&display=swap" rel="stylesheet">
-    </head>
-
-    <body>
-        <header>
-            <h1>Character Creator</h1>
-            <nav>
-                <ul>
-                    <li><a href="./index.html">Home</a></li>
-                    <li><a href="./creator.html">Creator</a></li>
-                    <li><a href="./info.html">About the Game</a></li>
-                    <li><a href="./chara.html">How To</a></li>
-                    <li><a href="./login.html">Login</a></li>
-                </ul>
-            </nav>
-        </header>
-        <main>
-            <section id="boxes">
-                <div class="about">Character Creator Description Here.</div>
-            </section>
-        </main>
-        <footer>&copy; 2022</footer>
-        <script src="./assets/script/indextabs.js"></script>
-        <script src="./public/js/logout.js"></script>
-    </body>
-</html>
