--- index.html
@@ -1,83 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-    <head>
-        <meta charset="UTF-8">
-        <meta name="viewport" content="width=device-width, initial-scale=1.0">
-        <title>Player Creator</title>
-        <link rel="stylesheet" href="./assets/style/style.css"/>
-        <link rel="stylesheet" href="./assets/style/index.css"/>
-        <link rel="preconnect" href="https://fonts.googleapis.com">
-        <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-        <link href="https://fonts.googleapis.com/css2?family=Cinzel+Decorative&display=swap" rel="stylesheet">
-        <link href="https://fonts.googleapis.com/css2?family=Raleway:wght@500&display=swap" rel="stylesheet">
-        <link href="https://fonts.googleapis.com/css2?family=Fondamento&display=swap" rel="stylesheet">
-        <link href="https://fonts.googleapis.com/css2?family=Amita&display=swap" rel="stylesheet">
-    </head>
-
-    <body>
-        <header>
-            <h1>Character Creator</h1>
-            <nav>
-                <ul>
-                    <li><a href="./index.html">Home</a></li>
-                    <li><a href="./creator.html">Creator</a></li>
-                    <li><a href="./info.html">About the Game</a></li>
-                    <li><a href="./chara.html">How To</a></li>
-                    <li><a href="./login.html">Login</a></li>
-                </ul>
-            </nav>
-        </header>
-        <main>
-            <section id="boxes">
-                <div class="altpic">
-                    <img class="slideshow" src="./assets/images/slideshow1.png" alt="orc picture"/>
-                    <img class="slideshow" src="./assets/images/slideshow2.png" alt="dwarf picture"/>
-                    <img class="slideshow" src="./assets/images/slideshow3.png" alt="drow picture"/>
-                    <img class="slideshow" src="./assets/images/slideshow4.png" alt="dragonborn picture"/>
-                </div>
-                <div class="basicdesc">
-                    <div class="desc">
-                        <div id="racedesc" class="fulldesc">
-                            <h2>Races</h2>
-                            <p>Every character belongs to a race, one of the many intelligent humanoid specied in the D&D world. The most common player character races are dwarves, elves, halflings, and humans. Some races also have subraces such as mountain dwarf or wood elf. The race you choose contributes to your character's identity in an important way, establishing a general appearance and the natural talents gained from culture and ancestry.</p>
-                            <p>Your character's race grants particular racial traits, such as special senses, proficiency with certain wepons or tools, proficiency in one or more skills, or the ability to use minor spells. These traits sometimes dovetail with the capabilities of certain classes. For Example, the racial traits of lightfoot halflings make them exceptional rogues, and high elves tend to be powerful wizards. Sometimes playing against type can be fun too. Half-Orc paladins and mountain dwarf wizards, for example, cam be unusual but memorable characters.</p>
-                        </div>
-                        <div id="classdesc" class="fulldesc">
-                            <h2>Classes</h2>
-                            <p>Every adventurer is a member of a class. Class broadly describes a character's vocation, what special talents they posess, and the tactics they are most likely to employ when exploring a dungeon, fighting monsters, or engaging in a tense negotiation.</p>
-                            <p>Your character receives a number of benefits from your choice of class. Many of these benefits are class features - capabilities (including spellcasting) that set your character apart from members of other lasses. You also gain a number of proficiencies: armor, weapons, skills, saving throws, and sometimes tool. Your proficiencies define many of the things your character can do particularly well, from using certain weapons to telling a convincing lie.</p>
-                        </div>
-                        <div id="subclassdesc" class="fulldesc">
-                            <h2>Subclasses</h2>
-                            <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.</p>
-                        </div>
-                        <div id="statsdesc" class="fulldesc">
-                            <h2>Stats</h2>
-                            <p>The stats of your character is a bit of a catchall phrase, as it isn't just a single section to go through.</p>
-                            <p>Much of what your character does in the game depends on their abilities: Strength, Dexterity, Constitution, Intelligence, Wisdom, and Charisma. Each ability has a score which is a number you receive either from a set string of numbers or by rolling a twenty sided die (D20) and figuring our the modifier that goes with the score you rolled.</p>
-                        </div>
-                        <div id="equipdesc" class="fulldesc">
-                            <h2>Equipment</h2>
-                            <p>When you create your character, you receive equipment based on a combination of your class and background. Alternatively, you can start with a number of gold pieces based on your class and spend them on items in game. You decide how your character came by this starting equipment. It might have been an inheritance, or goods that the character purchased during their upbringing. You might have been equipped with a weapon, armor, and a backpack as part of military service. You ight even have stolent your gear. A weapon could be a family heirloom, passed down from generation to generation until your character finally took up the mantle and followed in an ancestor's adventurous footsteps. Or you could have fought against someone and pilfered their body of their equipment. Whatever you decide, your strength stats determine how much your character is able to carry and each piece of equipment has a weight to it.</p>
-                        </div>
-                    </div>
-                    <div class="tabs">
-                        <button class="desctab" onclick="openDesc(event, 'racedesc')" id="defaultOpen">Race</button>
-                        <button class="desctab" onclick="openDesc(event, 'classdesc')">Class</button>
-                        <button class="desctab" onclick="openDesc(event, 'subclassdesc')">Subclass</button>
-                        <button class="desctab" onclick="openDesc(event, 'statsdesc')">Stats</button>
-                        <button class="desctab" onclick="openDesc(event, 'equipdesc')">Equipment</button>
-                    </div>
-                </div>
-            </section>
-            <section id="boxes">
-                <div  class="about">
-                    Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.
-                </div>
-            </section>
-        </main>
-        <footer>&copy; 2022</footer>
-        <script src="./assets/script/indextabs.js"></script>
-        <script src="./public/js/logout.js"></script>
-    </body>
-</html>
